requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.0
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
